backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-community
faiss-cpu
sentence-transformers
pytesseract
pdf2image
sqlite-utils
pydantic
easyocr